# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ResidentVoice Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithBuilding"
        "401":
          description: Unauthenticated
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithBuilding"
        "401":
          description: Invalid credentials
...
  /auth/register-building:
    post:
      operationId: registerBuilding
      tags: [auth]
      summary: Register a new building with admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingRegistrationInput"
      responses:
        "201":
          description: Building registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithBuilding"
        "400":
          description: Validation error
        "409":
          description: Slug already taken
  /auth/register-resident:
    post:
      operationId: registerResident
      tags: [auth]
      summary: Register as a resident in a building
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidentRegistrationInput"
      responses:
        "201":
          description: Resident registered (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithBuilding"
        "400":
          description: Validation error
        "404":
          description: Building not found

  # ─── Buildings ───────────────────────────────────────────────────────────────
...
  /buildings/{buildingId}/announcements:
    get:
      operationId: getAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcements list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement (admin)
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /buildings/{buildingId}/announcements/{announcementId}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement (admin)
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: integer
        - name: announcementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Notifications ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    BuildingRegistrationInput:
      type: object
      required: [buildingName, address, city, country, numberOfUnits, numberOfFloors, buildingType, adminFullName, adminEmail, adminPhone, password, slug]
      properties:
        buildingName:
          type: string
        slug:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        numberOfFloors:
          type: integer
        numberOfUnits:
          type: integer
        buildingType:
          type: string
          enum: [high_rise, condo, apartment, gated_community, residential_tower, mixed]
        adminFullName:
          type: string
        adminEmail:
          type: string
        adminPhone:
          type: string
        password:
          type: string
        adminRole:
          type: string
    ResidentRegistrationInput:
      type: object
      required: [buildingSlug, fullName, email, password, apartmentNumber, floor, residentType]
      properties:
        buildingSlug:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        apartmentNumber:
          type: string
        floor:
          type: integer
        residentType:
          type: string
          enum: [owner, tenant, family_member, property_manager]
        inviteCode:
          type: string
    UserWithBuilding:
      type: object
      required: [id, fullName, email, role, verificationStatus]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, building_admin, moderator, resident_owner, resident_tenant]
        verificationStatus:
          type: string
          enum: [pending, verified, rejected, suspended]
        apartmentNumber:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        residentType:
          type: ["string", "null"]
        buildingId:
          type: ["integer", "null"]
        building:
          $ref: "#/components/schemas/Building"
          nullable: true
...
    Building:
      type: object
      required: [id, name, slug, city, country, status, plan, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: ["string", "null"]
        country:
          type: string
        timezone:
          type: ["string", "null"]
        numberOfUnits:
          type: ["integer", "null"]
        numberOfFloors:
          type: ["integer", "null"]
        buildingType:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, pending, suspended]
        plan:
          type: string
          enum: [free, basic, pro, enterprise]
        createdAt:
          type: string
...
    Announcement:
      type: object
      required: [id, buildingId, title, message, category, createdAt]
      properties:
        id:
          type: integer
        buildingId:
          type: integer
        title:
          type: string
        message:
          type: string
        category:
          type: string
          enum: [general, emergency, maintenance, meeting, financial, security]
        priority:
          type: ["string", "null"]
        publishAt:
          type: ["string", "null"]
        expireAt:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, message, category]
      properties:
        title:
          type: string
        message:
          type: string
        category:
          type: string
          enum: [general, emergency, maintenance, meeting, financial, security]
        priority:
          type: string
        publishAt:
          type: string
        expireAt:
          type: string

    # ─── Notifications ────────────────────────────────────────────────────────
...